+++ audio/audiobuffer.h
@@ -14,7 +14,7 @@ public slots:
     void stop();
     void restart(int bufferInMsecs, const QAudioFormat& format);
 signals:
-    void bufferReady(const QVector<float> data);
+    void bufferReady(const QVector<double> data);
 protected:
     qint64 readData(char *data, qint64 maxlen) override;
     qint64 writeData(const char *data, qint64 len) override;
